Manage user

+++ app/admin/controllers/UserDisplayController.php
@@ -94,43 +94,59 @@ class UserDisplayController extends BaseController
 
     public function updateInsta($id){
         if(isset($_POST['sb-insta'])){
+            $link_ins = $_POST['link_image'];
             $image_old = insta::findOne($id)->link;
             $target_dir = "./public/upload/insta/";
             $nameimage = time() . $_FILES["upload-avatar-input"]["name"];
             $target_file = $target_dir . $nameimage;
             $errors = [];
             $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
-            if (file_exists($target_file)) {
-                $errors[] = "Sorry, file already exists.";
-            }
 
-// Check file size
-            if ($_FILES["upload-avatar-input"]["size"] > 500000) {
-                $errors[]= "Sorry, your file is too large.";
-            }
 
-// Allow certain file formats
-            if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
-                && $imageFileType != "gif" ) {
-                $errors[] = "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
-            }
 
             if(count($errors) > 0){
                 redirect('errors', $errors, 'edit-insta/'.$id);
             }else {
-                if (move_uploaded_file($_FILES["upload-avatar-input"]["tmp_name"], $target_file)) {
-                    $result = insta::updatefind($id,[
-                        "link" => $nameimage,
-                    ]);
-
-                    if (file_exists('./public/upload/insta/'.$image_old)) {
-                        unlink('./public/upload/insta/'.$image_old);
+                if(!($_FILES['logo']['name'] == "")) {
+                    if (empty($_POST['link_image'])) {
+                        $errors[] = 'Không được để trống link insta';
                     }
+
+// Check file size
+                    if ($_FILES["upload-avatar-input"]["size"] > 500000) {
+                        $errors[]= "Sorry, your file is too large.";
+                    }
+
+// Allow certain file formats
+                    if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
+                        && $imageFileType != "gif" ) {
+                        $errors[] = "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
+                    }
+                    if (move_uploaded_file($_FILES["upload-avatar-input"]["tmp_name"], $target_file)) {
+                        $result = insta::updatefind($id,[
+                            "link" => $nameimage,
+                            "link_insta" => $link_ins
+                        ]);
+
+                        if (file_exists('./public/upload/insta/'.$image_old)) {
+                            unlink('./public/upload/insta/'.$image_old);
+                        }
+                        if ($result){
+                            redirect('success', "Cập nhật thành công!", 'edit-insta/'.$id);
+                        }
+                    }
+
+                }else {
+                     insta::updatefind($id,[
+                        "link" => $image_old,
+                        "link_insta" => $link_ins
+                    ]);
+                    redirect('success', "Cập nhật thành công!", 'edit-insta/'.$id);
+
                 }
 
-                if ($result){
-                    redirect('success', "Cập nhật thành công!", 'edit-insta/'.$id);
-                }
+
+
             }
 
         }
